moved the swagger servlet registration to its own component
This should avoid potential circular dependencies at startup as the dashboard tile does not depend on HttpService anymore.
